2.8
{
    "filename": "links-2.8.tar.gz",
    "md5": "34c67f980105643fc6cd63a14acc626f",
    "sha1": "a92cf0c203fc765ef46974f15b2f738f2aec6b00",
    "sha256": "5070a759af7f107ca4f9572833b8f086cd9f7c21ef5e1fce8482a2883a743c7a"
}

+++ language.c
@@ -78,7 +78,7 @@ unsigned char *get_text_translation(unsigned char *text, struct terminal *term)
 			l_opt.plain = 0;
 			l_opt.cp = charset;
 			conv_table = get_translation_table(current_lang_charset, charset);
-			trn = convert_string(conv_table, tt, strlen(cast_const_char tt), &l_opt);
+			trn = convert_string(conv_table, tt, (int)strlen(cast_const_char tt), &l_opt);
 			if (!strcmp(cast_const_char trn, cast_const_char tt)) {
 				mem_free(trn);
 				trn = tt;
